+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/legacy/pipedata/TsFilePipeData.java
@@ -17,14 +17,14 @@
  * under the License.
  *
  */
-package org.apache.iotdb.db.pipe.connector.lagacy.pipedata;
+package org.apache.iotdb.db.pipe.connector.legacy.pipedata;
 
 import org.apache.iotdb.commons.exception.IllegalPathException;
 import org.apache.iotdb.db.engine.modification.ModificationFile;
 import org.apache.iotdb.db.engine.storagegroup.TsFileResource;
-import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.load.ILoader;
-import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.load.TsFileLoader;
-import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncConstant;
+import org.apache.iotdb.db.pipe.connector.legacy.pipedata.load.ILoader;
+import org.apache.iotdb.db.pipe.connector.legacy.pipedata.load.TsFileLoader;
+import org.apache.iotdb.db.pipe.connector.legacy.utils.SyncConstant;
 import org.apache.iotdb.tsfile.utils.ReadWriteIOUtils;
 
 import org.slf4j.Logger;
