+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/legacy/pipedata/queue/BufferedPipeDataQueue.java
@@ -16,14 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.pipe.connector.lagacy.pipedata.queue;
+package org.apache.iotdb.db.pipe.connector.legacy.pipedata.queue;
 
 import org.apache.iotdb.commons.exception.IllegalPathException;
 import org.apache.iotdb.commons.utils.FileUtils;
-import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.PipeData;
-import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.TsFilePipeData;
-import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncConstant;
-import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncPathUtil;
+import org.apache.iotdb.db.pipe.connector.legacy.pipedata.PipeData;
+import org.apache.iotdb.db.pipe.connector.legacy.pipedata.TsFilePipeData;
+import org.apache.iotdb.db.pipe.connector.legacy.utils.SyncConstant;
+import org.apache.iotdb.db.pipe.connector.legacy.utils.SyncPathUtil;
 import org.apache.iotdb.tsfile.exception.NotImplementedException;
 
 import org.slf4j.Logger;
