+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/legacy/pipedata/queue/PipeDataQueue.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,13 +16,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.pipe.connector.lagacy.exception;
+package org.apache.iotdb.db.pipe.connector.legacy.pipedata.queue;
 
-import org.apache.iotdb.pipe.api.exception.PipeException;
+import org.apache.iotdb.db.pipe.connector.legacy.pipedata.PipeData;
 
-public class SyncDataLoadException extends PipeException {
+import java.util.List;
 
-  public SyncDataLoadException(String message) {
-    super(message);
-  }
+public interface PipeDataQueue {
+  boolean offer(PipeData data);
+
+  List<PipeData> pull(long serialNumber);
+
+  PipeData take() throws InterruptedException;
+
+  void commit();
+
+  void commit(long serialNumber);
+
+  boolean isEmpty();
+
+  void close();
+
+  void clear();
 }
