+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/legacy/utils/SyncConstant.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.pipe.connector.lagacy.utils;
+package org.apache.iotdb.db.pipe.connector.legacy.utils;
 
 public class SyncConstant {
   /** common */
